Fix, improve and rename sire.sh
This fixes and improves the sire.sh script used for extending the QWAT
database with the ch_vd_sire extension.

The script is also renamed to sire_init.sh, for consistency with
qwat_init.sh.

This is how to initialize SIRE on a QWAT database created with
qwat_init.sh:

    ./extensions/ch_vd_sire/sire_init.sh -p qwat -s 21781 -d